extract the nxm protocol handler out of the nexus_integration entry
Moves the nxm:// link callback, the download resolver and the free-user
download queue into nxmProtocol.ts, and passes the Nexus connection in as a
parameter instead of reading it off a module-level binding the entry point
assigns during init. The handler can now be constructed in a test.

Drops three pieces of dead code the move exposed:
- makeNXMProtocol's onAwaitLink parameter, never called in the body
- the Bluebird onCancel handler in freeUserDownload; the only consumer
  (IPCDownloadAdapter.#resolve) awaits the handler as a plain promise, so
  nothing ever cancels it
- IDLQueueItem.canceled, only ever written
